"Fix" language file handling
The strings file has sprintf() format specifiers for the 16-bit MS-DOS
version. Since every parameter is implicitly converted to an int, this
isn't an issue for chars, shorts, int16_t, or uint16_t (the old int
and unsigned). However, some of the format specifiers want longs,
and those values are no longer a long type (they're int32_t). This
caused various numbers to be displayed incorrectly even though the
game used them correctly.

So, I went through and added casts to the type specified in the
strings file to every snprintf() that uses numbers (or chars).
This meant looking up the macro in mstrings.h, then looking at the
referenced line in string.txt... a lot.

Anyway, looks good now.

+++ src/doors/clans-src/src/game.c
@@ -92,7 +92,7 @@ void Game_Settings(void)
 	snprintf(szString, sizeof(szString), ST_GSETTINGS1, Game.Data.EliminationMode ? "On" : "Off");
 	rputs(szString);
 	*/
-	snprintf(szString, sizeof(szString), ST_GSETTINGS2, Game.Data.MaxPermanentMembers);
+	snprintf(szString, sizeof(szString), ST_GSETTINGS2, (int)Game.Data.MaxPermanentMembers);
 	rputs(szString);
 
 	if (Game.Data.InterBBS) {
@@ -103,11 +103,11 @@ void Game_Settings(void)
 	snprintf(szString, sizeof(szString), ST_GSETTINGS4, Game.Data.ClanEmpires ? "allowed" : "disabled");
 	rputs(szString);
 
-	snprintf(szString, sizeof(szString), ST_GSETTINGS5, Game.Data.MineFights);
+	snprintf(szString, sizeof(szString), ST_GSETTINGS5, (int)Game.Data.MineFights);
 	rputs(szString);
-	snprintf(szString, sizeof(szString), ST_GSETTINGS6, Game.Data.ClanFights);
+	snprintf(szString, sizeof(szString), ST_GSETTINGS6, (int)Game.Data.ClanFights);
 	rputs(szString);
-	snprintf(szString, sizeof(szString), ST_GSETTINGS7, Game.Data.DaysOfProtection);
+	snprintf(szString, sizeof(szString), ST_GSETTINGS7, (int)Game.Data.DaysOfProtection);
 	rputs(szString);
 	door_pause();
 }
